Extract file: octree_interesection.py
