+++ napari/layers/image/experimental/__init__.py
@@ -1,0 +1,1 @@
+from .octree_intersection import OctreeIntersection
